Refactor to have a separate bundle_utils.py
